requirements.txt
panel
crewai
python-dotenv
langchain
langchain-ollama
